+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to
 * Implemented `EthernetServer::availableForWrite()` as the minimum availability
   of all the connections, or zero if there's no connections.
 * New `AppWithListenersTemplate` example.
+* Added `EthernetClass::operator bool()` for testing whether Ethernet
+  is initialized.
 
 ### Changed
 * Changed `EthernetUDP::flush()` to be a no-op.